Move context extension to context extensions file
